Refactored script into more logical units
